requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-embed-video==1.1.2
djangorestframework==3.8.2
django-simple-captcha==0.5.9
factory-boy==2.11.1
Faker==0.8.17
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.5
requests==2.19.1
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.23
-e git+https://github.com/Tsuribori/django_rest_captcha_validator.git#egg=rest_validator
